Code improvements, SQL storage.
- Removed the Windows reporter.
- Merged ConcreteSketch and ConcreteSketchUncached into a single class
called Sketch.
- Cleaned up SQL storage. I don't recall why we used Derby but I've now
changed this class to use H2. I checked the licenses for both Derby and
H2 since I thought that may be why we switched but it turns out they're
both compatible with GPL. Since our documentation mentions H2 and we
also have traces using this, I don't think there should be any problem.
I think it is likely that I may have used Derby to test the stored
procedures (as indicated by some comments in the code).
- Fixed the erroneous link in the SQL documentation. There was an error
in one of the wiki links that may have occurred during the
GoogleCode-->GitHub migration. The link referred to under "Viewing the
H2 SQL database" in the ProvBench Traces wiki page now correctly points
to a tool for viewing the database.
- Cleaned up the custom vertex types.
- Renamed QueryTool to BatchTool.
- Updated the H2 SQL library dependency.
- A number of miscellaneous / minor code updates.