updated files to be filters
regex is divided into three parts. (part1) | (part2) | (part3). part1 consists of multiple patterns where the string starts with the patterns in part1, part2 consists of multiple patterns where the string contains any of the patterns in part2, and part3 consists of multiple patterns where the string ends with patterns in part3.