adds Kafka storage module
This storage module will serializes data into CDM v0.6 format using
Avro, and publish using Kafka.

There are some remaining items that need to be properly mapped into
CDM. Some of it is a matter of coding, and some is because of a
mismatch between the current data models (TRACE/SPADE vs CDM). Many
are noted in the code with 'XXX' marked comments.

We assume the elements (vertices and edges) received are in the
OPM syntax, SPADE's native format. The TC Common Data Model (CDM)
includes a properties field in all model elements for including
aribitrary key-value pairs for SPADE element annotations that do
not directly map to the CDM. NOTE: For these items, we use the Prov
element annotation key as the CDM properties' key, even though the
input from SPADE's reporter is in OPM syntax.

Syntax to add Kafka storage is as follows. All parameters are optional
to override defaults.

  add Kafka KafkaServer=<host:port> KafkaTopic=<topic> KafkaProducerID=<label ID> SchemaFilename=<filename>

Default parameter values:
  KafkaServer = localhost:9092
  KafkaTopic = trace-topic
  KafkaProducerID = trace-producer
  SchemaFilename = SPADE_ROOT/cfg/TCCDMDatum.avsc