Audit reporter:
Handled the case where a process which has exited (i.e. state cleared) is referred in a unit dependency record later.
Handling that by clearing the process state when a fork/clone happens on that pid again.
If this is not done then the process vertex used for dependency edge would not have any information that was needed and was only in the process maps that we keep.