CDM storage: Updated session number logic
If 'session' argument specified then that is used as session number. If not specified in arguments then file 'cfg/spade.storage.CDM.out' is read for getting the session number. If file not present or doesn't contain a key-value pair where key is 'lastSession' then session defaults to '0'. If session number found in file then it is incremented by one and that is used for the current CDM storage. Lastly, the used session number is written out to the mentioned file where key is 'lastSession'.