Audit: Added 'ptrace' syscall handling
The subset of 'ptrace' syscalls that are handled:
1) If the data or execution of the tracee is modified.
2) If the tracee process has been seen before.

OPM updates:
A) 'WasTriggeredBy' edge with operation 'ptrace' from the tracee process to the tracer process.
B) Annotation 'request' added on edge which has the value of the 'request' argument in the 'ptrace' syscall. Look at 'ptrace' manpage for list of possible values.